Reimplemented toplevel i/o using poor man's non-blocking instead of Glib. Toplevel working on Windows !
There are several reasons for this: after trying to use Glib's io, it
turns out that even though it tries to be portable it has different
functions / incomplete features on windows. Taking into account that
lablgtk adds another layer of the same kind of incompatibilities, and
that non-blocking on windows is quite hard (the previous version
did'nt work at all), we're much better off rewriting that the
old-fashioned way, with a reading thread (anyway, Glib seems to be
doing stuff like that behind the scenes...).

This version polls for input from the reading thread using a timeout
every few milliseconds: since the data is already in the program,
this is intended to be improved by propagating a Gdk event from the
reading thread.