A fix for some type inference problems
The inferred signature of `start` funcions in top.ml and topUi.ml
was less general than the declared one.

I've update the `.mli` to use unit instead of a polymorphic type
variable and add this change for an argument
```Ocaml
  let schedule f = .. (* return an id *)
  in
  let schedule f = ignore @@ schedule f (* returns unit *)
  in
```

The rest of the PR is just fixing dune warnings about unused
variables.